8382614: Consolidate implementation of current_stack_base_and_size on BSD
Apart from a workaround for some OSX versions on x86_64 architectures,
the implementation is identical for all architectures. This patch moves
it to the base os_bsd.cpp file so it's included regardless of target
architecture.

Sponsored by: The FreeBSD Foundation
